+++ Changelog.md
@@ -2,6 +2,8 @@
 
 ## Release 13
 
+* Added `knGetAppSdk` for getting the app's target SDK and `knGetDeviceSdk` for getting the device's android version (sdk level)
+* Added `knJavaCommand` for calling smash hit's java commands from lua
 * Added `knGetDeviceHz` and `knSetFrameRate`
 * Removed `knEnableReloading`; you can now just `knReload()` without enabling it
 * Fix a bug where the game would crash if using `knHttpRelease()` explicitly